online_inference/requirements.txt
pandas==1.5.1
fastapi==0.87.0
uvicorn==0.20.0
google-api-python-client==2.65.0